# openapi.yaml
openapi: 3.0.3
info:
  title: Provenact Control API
  version: 0.1.0
  description: Temporary control-plane API scaffold for Provenact.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/hash/sha256:
    post:
      summary: Compute prefixed sha256 digest for a payload string
      operationId: postSha256Hash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashRequest"
      responses:
        "200":
          description: Digest generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashResponse"
  /v1/verify/manifest:
    post:
      summary: Parse and validate a manifest, optionally enforce a policy capability ceiling
      operationId: postVerifyManifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyManifestRequest"
      responses:
        "200":
          description: Manifest accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyManifestResponse"
        "400":
          description: Invalid manifest or policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/verify/receipt:
    post:
      summary: Parse and verify an execution receipt hash
      operationId: postVerifyReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyReceiptRequest"
      responses:
        "200":
          description: Receipt accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyReceiptResponse"
        "400":
          description: Invalid receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/contexts:
    get:
      summary: List contexts for the authenticated owner
      operationId: getContexts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [starting, running, stopped, failed]
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Context list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListContextsResponse"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a context for the authenticated owner
      operationId: postContext
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContextRequest"
      responses:
        "200":
          description: Context created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateContextResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Package version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HashRequest:
      type: object
      required: [payload]
      properties:
        payload:
          type: string
          example: hello
    HashResponse:
      type: object
      required: [digest]
      properties:
        digest:
          type: string
          example: sha256:2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824
    VerifyManifestRequest:
      type: object
      required: [manifest]
      properties:
        manifest:
          oneOf:
            - $ref: "#/components/schemas/ManifestV0"
            - $ref: "#/components/schemas/ManifestV1Draft"
        policy:
          type: object
          nullable: true
          additionalProperties: true
    VerifyManifestResponse:
      type: object
      required: [schema_version, name, version, artifact, capability_ceiling_ok]
      properties:
        schema_version:
          type: string
          example: "0"
        name:
          type: string
          example: hello-skill
        version:
          type: string
          example: 0.1.0
        artifact:
          type: string
          example: sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
        capability_ceiling_ok:
          type: boolean
          example: true
    VerifyReceiptRequest:
      type: object
      required: [receipt]
      properties:
        receipt:
          oneOf:
            - $ref: "#/components/schemas/ExecutionReceiptV0"
            - $ref: "#/components/schemas/ExecutionReceiptV1Draft"
    VerifyReceiptResponse:
      type: object
      required: [schema_version, artifact, receipt_hash, valid]
      properties:
        schema_version:
          type: string
          example: "0"
        artifact:
          type: string
        receipt_hash:
          type: string
        valid:
          type: boolean
          example: true
    Capability:
      type: object
      required: [kind, value]
      properties:
        kind:
          type: string
        value:
          type: string
      additionalProperties: false
    EntryPointV1Draft:
      oneOf:
        - type: string
          minLength: 1
        - type: object
          required: [kind, path]
          properties:
            kind:
              type: string
              enum: [wasi-command, wasi-reactor]
            path:
              type: string
              minLength: 1
          additionalProperties: false
    JsonSchemaRefV1Draft:
      oneOf:
        - type: object
          additionalProperties: true
        - type: string
          format: uri
    ManifestV0:
      type: object
      required: [name, version, entrypoint, artifact, capabilities, signers]
      properties:
        name:
          type: string
        version:
          type: string
        entrypoint:
          type: string
        artifact:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/Capability"
        signers:
          type: array
          items:
            type: string
      additionalProperties: false
    ManifestV1Draft:
      type: object
      required:
        - schema_version
        - id
        - version
        - entrypoint
        - artifact
        - inputs_schema
        - outputs_schema
        - capabilities
        - signers
      properties:
        schema_version:
          type: string
          const: "1.0.0-draft"
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
        entrypoint:
          $ref: "#/components/schemas/EntryPointV1Draft"
        artifact:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        inputs_schema:
          $ref: "#/components/schemas/JsonSchemaRefV1Draft"
        outputs_schema:
          $ref: "#/components/schemas/JsonSchemaRefV1Draft"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/Capability"
        signers:
          type: array
          items:
            type: string
            minLength: 1
        provenance:
          type: object
          additionalProperties: true
        compatibility:
          type: object
          properties:
            runtime_profiles:
              type: array
              items:
                type: string
                minLength: 1
            adapter_profiles:
              type: array
              items:
                type: string
                minLength: 1
          additionalProperties: false
      additionalProperties: false
    ExecutionReceiptV0:
      type: object
      required: [artifact, inputs_hash, outputs_hash, caps_used, timestamp, receipt_hash]
      properties:
        artifact:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        inputs_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        outputs_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        caps_used:
          type: array
          items:
            type: string
        timestamp:
          type: integer
          minimum: 0
        receipt_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
      additionalProperties: false
    ExecutionResultV1Draft:
      type: object
      required: [status, code]
      properties:
        status:
          type: string
          enum: [success, failure]
        code:
          type: string
          minLength: 1
        message:
          type: string
      additionalProperties: false
    RuntimeV1Draft:
      type: object
      required: [name, version]
      properties:
        name:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
        profile:
          type: string
          minLength: 1
      additionalProperties: false
    AttestationV1Draft:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
      additionalProperties: false
    ExecutionReceiptV1Draft:
      type: object
      required:
        - schema_version
        - artifact
        - manifest_hash
        - policy_hash
        - inputs_hash
        - outputs_hash
        - caps_requested
        - caps_granted
        - caps_used
        - result
        - runtime
        - started_at
        - finished_at
        - receipt_hash
      properties:
        schema_version:
          type: string
          const: "1.0.0-draft"
        artifact:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        manifest_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        policy_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        inputs_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        outputs_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        caps_requested:
          type: array
          items:
            type: string
            minLength: 1
        caps_granted:
          type: array
          items:
            type: string
            minLength: 1
        caps_used:
          type: array
          items:
            type: string
            minLength: 1
        result:
          $ref: "#/components/schemas/ExecutionResultV1Draft"
        runtime:
          $ref: "#/components/schemas/RuntimeV1Draft"
        started_at:
          type: integer
          minimum: 0
        finished_at:
          type: integer
          minimum: 0
        attestations:
          type: array
          items:
            $ref: "#/components/schemas/AttestationV1Draft"
        receipt_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
      additionalProperties: false
...
    ContextSummary:
      type: object
      required: [id, status, region, started_at, last_activity]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [starting, running, stopped, failed]
        region:
          type: string
        started_at:
          type: string
        ended_at:
          type: string
          nullable: true
        package:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
        last_activity:
          type: string
    ListContextsResponse:
      type: object
      required: [contexts]
      properties:
        contexts:
          type: array
          items:
            $ref: "#/components/schemas/ContextSummary"
    CreateContextRequest:
      type: object
      required: [status, region]
      properties:
        status:
          type: string
          enum: [starting, running, stopped, failed]
        region:
          type: string
          example: us-east-1
        package:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
    CreateContextResponse:
      type: object
      required: [context]
      properties:
        context:
          $ref: "#/components/schemas/ContextSummary"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string